Replace temperature system Before: integer precision temperature system After: double precision energy based temperature system

+++ PixelEngine/Chunk.hpp
@@ -27,7 +27,7 @@
 typedef struct {
     Uint16 index;
     Uint32 color;
-    int32_t temperature;
+    double temperature;
 } MaterialInstanceData;
 
 class Chunk {
